(기능) 결제 방법에 최종 결제할 금액을 계산 출력 하는 기능 구현
  - 최종결제 방법은
    같은 메뉴에 한해서 10개 단위로 10000원씩 할인된다.
    현금 결제시 5퍼센트 할인된 가격으로 결제 한다.

+++ src/main/java/view/OutputView.java
@@ -66,7 +66,7 @@ public class OutputView {
     }
 
     public static void moneyPayment(int allPayment) {
-        System.out.println("최종 결제할 금액 " + allPayment * 0.95);
+        System.out.println("최종 결제할 금액 " + (int)(allPayment * 0.95));
     }
 
     public static void cardPayMent(int allPayment) {
